areaoftriangle

+++ src/com/company/AreaOfCircle.java
@@ -6,7 +6,7 @@ public class AreaOfCircle {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("Enter the radius: ");
-        double radius = in.nextFloat();
+        double radius = in.nextDouble();
         double area =  (3.14*radius*radius);
         System.out.println("Area of Circle is: "+ area);
     }
